requirements .txt
cv2
numpy
telegram.ext
urllib